# tmp/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Amor Miau API Documentation",
    "contact": {
      "email": "leo.brsouza@gmail.com"
    },
    "description": "Amor Miau is an animal welfare NGO. learn more: https://bit.ly/githubAmorMiau",
    "license": {
      "name": "MIT license",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "amormiau-backend.herokuapp.com",
  "basePath": "/",
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "x-access-token"
    }
  },
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/admin/registerCat": {
      "post": {
        "summary": "Register Cat",
        "tags": ["Cat"],
        "operationId": "RegisterCat",
        "deprecated": false,
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Cat's name",
            "required": true,
            "type": "string"
          },
          {
            "name": "birthDate",
            "in": "formData",
            "description": "Cat's birth date",
            "required": true,
            "type": "string"
          },
          {
            "name": "weight",
            "in": "formData",
            "description": "Cat's weight",
            "required": false,
            "type": "string"
          },
          {
            "name": "sterilized",
            "in": "formData",
            "description": "Cat's sterilized",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "specialCat",
            "in": "formData",
            "description": "Special cat",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "available",
            "in": "formData",
            "description": "Cat's available",
            "required": true,
            "type": "boolean"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "Cat's description",
            "required": false,
            "type": "string"
          },
          {
            "name": "sociable",
            "in": "formData",
            "description": "Cat's sociable level",
            "required": false,
            "type": "number"
          },
          {
            "name": "playful",
            "in": "formData",
            "description": "Cat's playful level",
            "required": false,
            "type": "number"
          },
          {
            "name": "affectionate",
            "in": "formData",
            "description": "Cat's affectionate level",
            "required": false,
            "type": "number"
          },
          {
            "name": "images",
            "in": "formData",
            "description": "Cat's image",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/admin/cat/{catId}": {
      "get": {
        "summary": "Get cat by ID",
        "tags": ["Cat"],
        "operationId": "GetcatbyID",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update Cat",
        "tags": ["Cat"],
        "operationId": "UpdateCat",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateCatRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete Cat",
        "tags": ["Cat"],
        "operationId": "DeleteCat",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "UpdateCatRequest": {
      "title": "UpdateCatRequest",
      "example": {
        "name": "Meow",
        "birthDate": "03/02/2017",
        "weight": "6kg",
        "sterilized": true,
        "specialCat": true,
        "available": true
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "birthDate": {
          "type": "string"
        },
        "weight": {
          "type": "string"
        },
        "sterilized": {
          "type": "boolean"
        },
        "specialCat": {
          "type": "boolean"
        },
        "available": {
          "type": "boolean"
        },
        "sociable": {
          "type": "number"
        },
        "playful": {
          "type": "number"
        },
        "affectionate": {
          "type": "number"
        }
      },
      "required": ["name", "birthDate", "available"]
    },
...
  },
...
}